+++ test/test_helper.exs
@@ -1,4 +1,4 @@
-Postgrex.Types.define(EctoApp.PostgrexTypes, [Pgvector.Extensions.Vector] ++ Ecto.Adapters.Postgres.extensions(), [])
+Postgrex.Types.define(EctoApp.PostgrexTypes, Pgvector.extensions() ++ Ecto.Adapters.Postgres.extensions(), [])
 
 defmodule Repo do
   use Ecto.Repo,